+++ src/app/blog/[slug]/page.tsx
@@ -1,13 +1,16 @@
 import { MDXRemote } from 'next-mdx-remote/rsc'
 import { Suspense } from 'react'
+import Image from 'next/image'
 
 import { getArticleFile, getArticleInfo } from '@/api/blog/Articles'
 import { SOURCE } from '@/constants/static'
-import Image from 'next/image'
+import BlogClock from '@/constants/svg/BlogClock'
+
+import ArticleImage from './components/ArticleImage'
 import Carousel from './components/Carousel'
 
-import BlogClock from '@/constants/svg/BlogClock'
 import style from './style.module.css'
+
 export const GetPost = async (slug: string) => {
   return getArticleFile({ slug })
 }
@@ -25,9 +28,19 @@ const MDXPage = async ({ params }: IPostPage) => {
 
   const overrideComponents = {
     Carousel: Carousel,
+    Image: ArticleImage,
   }
 
-  return <MDXRemote source={content || ''} components={overrideComponents} />
+  console.log(content)
+
+  try {
+    // return <>сигма1</>
+    return <MDXRemote source={content || ''} components={overrideComponents} />
+  } catch (e) {
+    console.log(e)
+    return <>сигма</>
+
+  }
 }
 
 const ArticleHeader = async ({ params }: IPostPage) => {
@@ -37,7 +50,7 @@ const ArticleHeader = async ({ params }: IPostPage) => {
   return (
     <>
       {/* Image placeholder="blur" */}
-      <Image className="rounded-xl m-0 mb-3" width={1600} height={900}  src={`${SOURCE.url}posts/${articleInfo?.slug}/files/preview`} alt={'изображение'} />
+      <Image className="rounded-xl m-0 mb-3" width={1600} height={900} src={`${SOURCE.url}posts/${articleInfo?.slug}/files/preview`} alt={'изображение'} />
       <div className="flex my-4 text-[16px] font-medium items-center text-center">
         {date}
         <span className={`${style.slash} mx-1`} />
