+++ src/constants/static.tsx
@@ -11,3 +11,146 @@ export const SOURCE = {
 			: 'https://chat-boty.com/api/v1/',
 	// source_url: 'https://resource.chat-boty.com/',
 }
+
+
+export const textReviews = [
+  {
+    text: 'Начали сотрудничать с Никитой в январе 2020 года. За время своей работы Никита внедрил множество инструментов геймификации, настроил реферальную систему, которой нету ни у кого среди конкурентов, внедрял постоянно чат-ботов, о новых трендах в ВК мы узнавали первыми.\n\nСделал файл со 100 полезными сервисами, и рассказал не очевидные функции сенлера и других сервисов.\n\nНикита показал себя, как опытный и исполнительный партнер. Все всегда делалось вовремя и в темпе.',
+    title: 'о новых трендах в ВК мы узнавали первыми.',
+    company: 'Дайбобру',
+    name: 'Ярослав Миньков',
+    authorLink: 'https://vk.com/yr_minkov',
+    avatar: '/yarik.webp',
+  },
+  {
+    text: 'Неоднократно обращалась\nк Никите по технической части\nВк: сенлер, автопилот,\nигрофикация. Всегда очень\nдовольна результатом. Очень\nтерпеливо отвечает на кучу\nвопросов. Объясняет быстро,\nчётко и доходчиво. Когда нужно выполняет сам, так же быстро\nи с учётом всех моих пожеланий.\nВсегда на связи.',
+    title: 'быстро, четко,\n доходчиво',
+    company: 'додо пицца',
+    name: 'Елена Попова',
+    authorLink: 'https://vk.com/id56103386',
+    avatar: '/otzev_1.webp',
+  },
+  {
+    text: 'Консалтился у Никиты по онлайн школам в мае\n\nНикита Круто очень рассказал про автопилот, чат боты, новый подход к Запускам, воронкам\n\nТо что творит с автопилотом, хитрыми фишками ВК и тд это жара👍👍🔥\n\nЭто маркетинг нового поколения🚀\n\nПодход огненный)\nРекомендую',
+    title: 'Это маркетинг нового поколения',
+    company: 'Продюсер',
+    name: 'Евгений Ренуа',
+    authorLink: 'https://vk.com/evgeny.renua',
+    avatar: '/reuna.webp',
+  },
+  {
+    text: 'Сегодня в чате поддержки АП я задал вопрос про VK donut и реализацию уведомлений по ним и Никита  пришёл ко мне на помощь\n\nБолее того, я веду сообщество в вк донат с первого дня создания подписной стены и я всегда мечтал (именно мечтал, по-другому не скажешь) настроить связь vk donut с АП и Никита все сделал очень быстро.\n\nНикита все сделал очень быстро! Спасибо огромное, Никита! Буду обращаться ещё!',
+    title: 'Мечтал настроить связь vk donut с Автопилотом, Никита всё сделал быстро ',
+    company: 'Преподаватель английского',
+    name: 'Фаил Файзуллин',
+    authorLink: 'https://vk.com/frfayzullin',
+    avatar: '/frfayzullin.webp',
+  },
+  {
+    text: 'Обратился к Никите в 2022 году для создания клуба по подписке, сначала всё было сделано в ВК, чуть позже Никита всё перенёс на WB.\n\nВсё работает до сих пор с минимумом моего участия. По некоторым техническим вопросам я пишу Никите и он всё оперативно реализует. ',
+    title: 'Клуб по подписке работает как часы уже 2 года',
+    company: 'Seller на WB',
+    name: 'Егор Опарин',
+    authorLink: 'https://t.me/egor_oparin',
+    avatar: '/oparin.webp',
+  },
+  {
+    text: 'Работали с Никитой с конца февраля, помогал смыслами с воронкой, дожимами. Делал быстро всю техническую часть по ботам в телеграм и инстаграм.\n\nСделал сайт на креатиум быстро, не нужно было искать 100500 специалистов. Помог снизить комиссию по эквайрингу, остались довольны работой',
+    title: 'Стабильно делали много рассылок и ботов, Никита со всем справлялся',
+    company: 'Продюсер',
+    name: 'Алёна Мелниченко',
+    authorLink: 'https://t.me/alena_melnic',
+    avatar: '/melnic.webp',
+  },
+	{
+    text: 'Работаем с Никитой с начала 2023 года, работает с базой в ВК, ТГ, Инсте и e-mail, также пробовали SMS\n\nСейчас подключили WhatsApp и работа идёт ежедневно по всем направлениям. Радует, что Никита всегда на связи, показал много инструментов в ботах, которые мы не использовали на 100%.\n\nТакже ведёт карты и работает с репутацией, всем довольны. Продолжаем работать',
+    title: 'Работаем по всем фронтам: TG, ВК, Инста, WhatsApp и E-mail уже второй год',
+    company: 'РОМ в CHUKCHA',
+    name: 'Геогрий Тарасов',
+    authorLink: 'https://vk.com/1gt3a5l',
+    avatar: '/chukcha.webp',
+  },
+	{
+    text: 'Работаю с Никитой чуть-ли не с 2019 года, делал автоматизацию и группы ВК, подключал и настраивал виджеты, сейчас делает оперативно больших ботов в ТГ, настраивает ботов в инсте.\n\nПостоянно на связи, что очень важно для меня, быстро вносит правки. Доволен нашим сотрудничеством',
+    title: 'Работаю с Никитой с 2019 года, быстро вносит правки.',
+    company: 'Продюсер',
+    name: 'Владимир Несмеянов',
+    authorLink: 'https://t.me/nesmeyanov555',
+    avatar: '/nesmeyanov.webp',
+  },
+	{
+    text: 'Работаю с Никитой по рекомендации другого продюсера, мне нравится, что с ним можно созвониться и собрать на месте бота, не нужно ждать неделями всё, также правки все вносились оперативно.\n\nПо ботам ВК, ТГ и Инст - пишу Никите, он может сделать всё под мои запросы',
+    title: 'Делает ботов на месте, пока общаемся, по ботам только к нему',
+    company: 'Продюсер',
+    name: 'Евгений Рукин',
+    authorLink: 'https://vk.com/evgen_rukin',
+    avatar: '/rukin.webp',
+  },
+	{
+    text: 'Никита помог с задачами по интеграции Геткурса с автопилотом. Хороший парень, всегда на связи, делает хорошо, быстро, и с пониманием дела. Рекомендую 😎',
+    title: 'всегда на связи, делает быстро и с пониманием дела',
+    company: 'Алхимия Дизайна',
+    name: 'Иван Завалко',
+    authorLink: 'https://vk.com/vzavalko',
+    avatar: '/zavalko.webp',
+  },
+]
+
+export const videoReviews = [
+  {
+    // href: 'https://www.youtube.com/embed/G0fHsOQWG5g',
+		id:'456239071',
+		oid:'-165380047',
+    vertical: '/video/afoninV.webp',
+    horizontal: '/video/afoninH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/sW521Rh11Wo',
+		id:'456239053',
+		oid:'-165380047',
+    vertical: '/video/milovV.webp',
+    horizontal: '/video/milovH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/KKhLx4ETy0M',
+		id:'456239069',
+		oid:'-165380047',
+    vertical: '/video/ivanovV.webp',
+    horizontal: '/video/ivanovH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/h74TsLk3xHA',
+		id:'456239057',
+		oid:'-165380047',
+    vertical: '/video/zhamkevichV.webp',
+    horizontal: '/video/zhamkevichH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/tILvQMkBqbw',
+		id:'456239054',
+		oid:'-165380047',
+    vertical: '/video/emilV.webp',
+    horizontal: '/video/emilH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/LHmSlMMLcB8',
+		id:'456239065',
+		oid:'-165380047',
+    vertical: '/video/yanbukhtinV.webp',
+    horizontal: '/video/yanbukhtinH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/uUHp-eNZnkM',
+		id:'456239056',
+		oid:'-165380047',
+    vertical: '/video/nasirovV.webp',
+    horizontal: '/video/nasirovH.webp',
+  },
+  {
+    // href: 'https://www.youtube.com/embed/tUkagWy8nIQ',
+		id:'456239058',
+		oid:'-165380047',
+    vertical: '/video/blackV.webp',
+    horizontal: '/video/blackH.webp',
+  },
+]
