+++ src/modules/landing/FiveScreen/index.tsx
@@ -1,12 +1,6 @@
-import { Manrope } from 'next/font/google'
-import { TextCarousel, VideoCarousel } from './Carousel'
+import { TextCarousel } from './modules/TextCarousel'
+import { VideoCarousel } from './modules/VideoCarousel'
 import styles from './style.module.css'
-
-const nunitoFont = Manrope({
-  // weight: '700',
-  subsets: ['latin'],
-  // variable: '--font-poppins',
-})
 
 export const FiveScreen = () => {
   return (
