+++ src/modules/landing/FiveScreen/modules/VideoModal.tsx
@@ -1,12 +1,28 @@
 import { PopUp } from '@/components/ui/PopUp'
 import CloseMainPopUp from '@/constants/svg/CloseMainPopUp'
 
-export const VideoModal = ({ show, src, closePopup, image }: { show: boolean; src: string; closePopup: () => void; image:string }) => {
+export const VideoModal = ({ show, oid, id, closePopup, image }: { show: boolean; oid: string; id: string; closePopup: () => void; image: string }) => {
   return (
-    <PopUp show={show} className=''>
+    <PopUp show={show} className="">
       <div className={'flex flex-col items-center '}>
         <CloseMainPopUp className={'fixed'} style={{ right: '20px', top: '20px' }} onClick={closePopup} />
+        {/* <iframe
+          src={`https://vk.com/video_ext.php?oid=${oid}&id=${id}&hd=2`}
+          width="853"
+          height="480"
+          allow="autoplay; encrypted-media; fullscreen; picture-in-picture; screen-wake-lock;"
+          frameBorder="0"
+          allowFullScreen
+        ></iframe> */}
         <iframe
+          src={`https://vk.com/video_ext.php?oid=${oid}&id=${id}&hd=2&autoplay=1`}
+          width="853"
+          height="480"
+          allow="autoplay; encrypted-media; fullscreen; picture-in-picture; screen-wake-lock;"
+          frameBorder="0"
+          allowFullScreen
+        ></iframe>
+        {/* <iframe
           className='max-w-[840px] max-h-[500px] w-[80vw] h-[50vw]'
           width="100%"
           height="100%"
@@ -23,7 +39,7 @@ export const VideoModal = ({ show, src, closePopup, image }: { show: boolean; sr
           allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
           allowFullScreen
           title="Demo video"
-        />
+        /> */}
       </div>
     </PopUp>
   )
