+++ src/modules/landing/FiveScreen/modules/components/style.module.css
@@ -1,0 +1,19 @@
+
+.cardText {
+  color: var(--color-additional-3);
+}
+
+.cardTitle {
+  color: var(--color-additional);
+  font-family: 'Igra Sans', sans-serif;
+}
+
+.cardName {
+  color: var(--color-additional-3);
+}
+
+.avatar {
+  width: 50px;
+  height: 50px;
+  background-color: var(--color-additional-3);
+}
