+++ src/modules/landing/FiveScreen/modules/components/ui/Splitter.tsx
@@ -1,0 +1,3 @@
+import styles from './style.module.css'
+
+export const Splitter = () => <div className={`${styles.splitter} mb-10`} />
