+++ src/modules/landing/FiveScreen/modules/style.module.css
@@ -1,0 +1,66 @@
+.screen {
+  background-color: #fff;
+  border-radius: 35px;
+  background-size: 50% auto;
+}
+
+.title {
+  color: var(--color-accent);
+}
+
+.card {
+  background-color: var(--color-additional-2);
+}
+
+.cardText {
+  color: var(--color-additional-3);
+}
+
+.cardTitle {
+  color: var(--color-additional);
+  font-family: 'Igra Sans', sans-serif;
+}
+
+.cardName {
+  color: var(--color-additional-3);
+}
+
+.splitter {
+  width: 100%;
+  height: 1px;
+  background-color: var(--color-additional-3);
+}
+
+.avatar {
+  width: 50px;
+  height: 50px;
+  background-color: var(--color-additional-3);
+}
+
+.carousel {
+  /* width: 100vw; */
+  display: flex;
+  overflow-x: scroll;
+
+}
+
+.gallery {
+  /* width: 100%; */
+  display: flex;
+  grid-template-columns: auto auto auto;
+  gap: 20px;
+  padding: 10px;
+  flex: none;
+}
+
+.carousel::-webkit-scrollbar {
+  display: none;
+}
+
+.vector {
+  transition: all .2s linear;
+}
+
+.vector:hover {
+  fill: var(--color-accent-2);
+}
